Add test 'onSubmit handler works as expected'
