+++ src/containers/users/tests/UserForm.test.jsx
@@ -5,7 +5,6 @@ import { shallow, mount } from 'enzyme';
 import toJson from 'enzyme-to-json';
 import { UserForm } from '../UserForm';
 import * as fixtures from '../../../store/users/tests/fixtures';
-import response1 from './fixtures';
 
 global.fetch = require('jest-fetch-mock');
 
@@ -32,42 +31,5 @@ describe('containers/users/UserForm', () => {
     );
     expect(toJson(wrapper)).toMatchSnapshot();
     wrapper.unmount();
-  });
-  it('onSubmit handler works as expected', () => {
-    const wrapper = mount(
-      <UserForm
-        formActionDispatch={function() {}}
-        initialData={fixtures.UserFormInitialData}
-        targetId={1337}
-      />
-    );
-
-    fetch.mockResponseOnce(JSON.stringify(response1), { status: 400 });
-
-    const values = {
-      confirmation: 'hello123',
-      email: 'kamau@gmail.com',
-      expertise: 'Intermediate',
-      first_name: 'kamau',
-      gender: 'male',
-      last_name: 'kahama',
-      national_id: '30090087',
-      ona_username: 'kahama',
-      password: 'hello123',
-      payment_number: '',
-      phone_number: '',
-      role: 'Admin'
-    };
-
-    const setSubmitting = jest.fn();
-    const setErrors = jest.fn();
-    const setStatus = jest.fn();
-
-    wrapper
-      .find('Formik')
-      .props()
-      .onSubmit(values, { setSubmitting, setErrors, setStatus });
-
-    expect(toJson(wrapper.find('form'))).toMatchSnapshot('formik');
   });
 });
